+++ docs/source/changelog.rst
@@ -2,7 +2,26 @@
 Changelog
 =========
 
-- 
+1.0.1
+^^^^^
+
+November 14, 2020
+
+- Added lets/generate/tls/certificate
+- Added lets/__images__/local/socat
+- Added lets/proxy/forward/http
+- Added lets/proxy/forward/socks5
+- Added lets/proxy/forward/openvpn/http
+- Added lets/proxy/forward/openvpn/socks5
+- Added lets/proxy/forward/tor/http
+- Added lets/proxy/forward/tor/socks
+- Added lets/proxy/reverse/tor/service
+- Added lets/proxy/reverse/ngrok/http
+- Added lets/redirect/socat
+- Added lets/redirect/tcp/tcp
+- Added lets/redirect/tcp/tls
+- Added lets/redirect/tls/tcp
+
 
 1.0.0
 ^^^^^